So much work to inline svg elements with handlebars
Had to end up building a custom loader that parses the output from
underscore-loader-template and inlines the svgs before evaluating the
function and passing to handlebars-loader

I played with so much crap during this to get it to work. Finally found
this solution that lets us use the same syntax for both static templates
and handlebars partials.